Rename SchemaRegisterForward to SchemaRegisterProcessor
