+++ src/main/resources/log4j.properties
@@ -7,7 +7,7 @@ log4j.logger.com.datafibers.service.DFWebUI=DEBUG
 log4j.logger.com.datafibers.service.DFDataProcessor=DEBUG
 log4j.logger.com.datafibers.processor.FlinkTransformProcessor=DEBUG
 log4j.logger.com.datafibers.processor.KafkaConnectProcessor=DEBUG
-log4j.logger.com.datafibers.processor.SchemaRegisterForward=DEBUG
+log4j.logger.com.datafibers.processor.SchemaRegisterProcessor=DEBUG
 log4j.logger.com.datafibers.util.HelpFunc=DEBUG
 log4j.logger.com.datafibers.util.CLIParser=INFO
